Build Windows version
